Refactor tests, correctly adding subscriptions still fails
