--- gridt/src/app/core/models/subscription-holder.model.ts
@@ -1,9 +1,0 @@
-import { Subscription } from 'rxjs';
-
-export class SubscriptionHolder {
-    protected subscriptions: Subscription[] = [];
-
-    cancelAllSubscriptions () {
-        this.subscriptions.map( subscription => subscription.unsubscribe() );
-    }
-}
