+++ gridt/src/app/movements/add-movement/add-movement.page.ts
@@ -4,7 +4,7 @@ import { LoadingController, AlertController } from '@ionic/angular';
 import { Router } from '@angular/router';
 import { ApiService } from '../../core/api.service';
 import { Movement } from '../../core/models/movement.model';
-import { SubscriptionHolder } from 'src/app/core/models/subscription-holder.model';
+import { SubscriptionHolder } from 'src/app/core/subscription-holder.extendableClass';
 
 @Component({
   selector: 'app-add-movement',
